+++ src/exceptions.py
@@ -1,0 +1,13 @@
+# Custom exception for a failed simulation
+
+
+class FailedSimulationException(Exception):
+    pass
+
+
+# Custom exception for a OSM error.
+
+
+class OSMErrorException(Exception):
+    pass
+
