+++ src/helper.py
@@ -1,0 +1,56 @@
+import os
+import logging
+import yaml
+from .aws_batch import AWSBatch
+from .docker_batch import DockerBatch
+
+def batch_factory(
+                  compute_environment=None,
+                  analysis_id=None,
+                  btap_image_name=None,
+                  nocache=True,
+                  os_version=None,
+                  btap_costing_branch=None,
+                  os_standards_branch=None,
+                  git_api_token=None):
+    batch = None
+    if compute_environment == 'aws_batch':
+        # create aws image, set up aws compute env and create workflow queue.
+
+        batch = AWSBatch( analysis_id=analysis_id,
+                          btap_image_name=btap_image_name,
+                          rebuild_image=nocache,
+                          git_api_token=git_api_token,
+                          os_version=os_version,
+                          btap_costing_branch=btap_costing_branch,
+                          os_standards_branch=os_standards_branch
+                          )
+        # Create batch queue on aws.
+        batch.setup()
+    elif compute_environment == 'local':
+        batch = DockerBatch(image_name=btap_image_name,
+                            git_api_token=git_api_token,
+                            os_standards_branch=os_standards_branch,
+                            btap_costing_branch=btap_costing_branch,
+                            os_version=os_version,
+                            nocache=nocache
+                            )
+        # Create batch queue on docker desktop.
+        batch.setup()
+    return batch
+
+
+# Helper method to load input.yml file into data structures required by btap_batch
+def load_btap_yml_file(analysis_config_file):
+    # Load Analysis File into variable
+    if not os.path.isfile(analysis_config_file):
+        logging.error(f"could not find analysis input file at {analysis_config_file}. Exiting")
+        exit(1)
+    # Open the yaml in analysis dict.
+    with open(analysis_config_file, 'r') as stream:
+        analysis = yaml.safe_load(stream)
+    # Store analysis config and building_options.
+    analysis_config = analysis[':analysis_configuration']
+    building_options = analysis[':building_options']
+    return analysis_config, building_options
+
